Link to the images in lesson 1
By the time a student gets to lesson 4, they might have forgotten where
they put the images in lesson 1. Or they might have skipped lesson 1
entirely.

In the latter case, we probably need some way to get all the files for
lesson 1 without actually doing it... But that requires a bit more
thought...