Move dots to the same line as `.sub()` calls
I think originally I did this with the dots at the end of the lines to
make it "more obvious" to the reader that the line wasn't finished (and
that the next line followed on from it).

I've seen a couple of people get confused by it though - they don't see
the dot, and instead do something like:

```
response.body = File.read("barking-permit-template.html")
  sub("DOG_NAME", dog_name)
  sub("DOG_AGE_HUMAN_YEARS", dog_age).
```

... which results in `object:Main has no method sub` or something like
that. Which is confusing.

I think it's a little bit easier to spot the dot if it's on the
beginning of the line.

Still confusing, but hey, it's coding. It's famously confusing.

At least this is a possible opportunity to show the difference between
top level methods and instance methods.