Replace therubyracer with mini_racer
Building this site requires a JS runtime, to be able to use
`autoprefixer-rails`. Previously this was provided by `therubyracer`,
however versions of `govuk_tech_docs` >= 2.0.12 use autoprefixer-rails
v9, which doesn't work with `therubyracer`. However, it does with with
`mini_racer`, an alternative bridge with the JS runtime, so this commit
swaps the one for the other.